Move `Meditor.vue` to components directory It's now part of the DLP view instead of being its own self-contained view
